Add notifications app
Added notifications app for sending notifications about sign-ups.
Currently email notifications about sign-up creations are supported.
The app should make it easy add to new notification types in the
future.

The actual messages for the notifications are be entered using the
admin UI. The messages can also contain sign-up specific values, the
user's name for example. Localization support for the messages is also
implemented, but currently only Finnish is in use.